Backwards incompatible: Require explicit command line actions
Ever since the first releases it bothered me that the collection of
metrics in /tmp/apache-manager.txt was enabled rather implicitly.

Also given recent developments the killing of workers couldn't be
disabled if a configuration file defined thresholds. This was caused by
a combination of the following circumstances:

- The command line interface was initially designed such that killing
  of workers was enabled when thresholds were given as command line
  options.

- Since then support for configuration files was added, and given the
  presence of a configuration file thresholds would always be enabled
  so killing would happen implicitly and unconditionally.

To solve all of this anti "explicit is better than implicit" magic I
decided to make a backwards incompatible change to the command line
interface, where both of the actions described above now need to be
requested using command line options.